// src/store/modules/profile/mutation-types.ts
export enum MutationTypes {
  ONBOARDING_STEP1_LOADING = "ONBOARDING_STEP1_LOADING",
  ONBOARDING_STEP1_LOADED = "ONBOARDING_STEP1_LOADED",
  ONBOARDING_STEP2_LOADING = "ONBOARDING_STEP2_LOADING",
  ONBOARDING_STEP2_LOADED = "ONBOARDING_STEP2_LOADED",
  ONBOARDING_STEP3_LOADING = "ONBOARDING_STEP3_LOADING",
  ONBOARDING_STEP3_LOADED = "ONBOARDING_STEP3_LOADED",
  ONBOARDING_STEP4_LOADING = "ONBOARDING_STEP4_LOADING",
  ONBOARDING_STEP4_LOADED = "ONBOARDING_STEP4_LOADED",
  ONBOARDING_STEP5_LOADING = "ONBOARDING_STEP5_LOADING",
  ONBOARDING_STEP5_LOADED = "ONBOARDING_STEP5_LOADED",
  ONBOARDING_STEP3_DATA_LOADING = "ONBOARDING_STEP3_DATA_LOADING",
  ONBOARDING_STEP3_DATA_LOADED = "ONBOARDING_STEP3_DATA_LOADED",
  ONBOARDING_STEP4_DATA_LOADING = "ONBOARDING_STEP4_DATA_LOADING",
  ONBOARDING_STEP4_DATA_LOADED = "ONBOARDING_STEP4_DATA_LOADED",
  ZIP_CITY_LOADING = "ZIP_CITY_LOADING",
  ZIP_CITY_LOADED = "ZIP_CITY_LOADED",
  UPLOAD_CONFIGURATIONS_LOADING = "UPLOAD_CONFIGURATIONS_LOADING",
  UPLOAD_CONFIGURATIONS_LOADED = "UPLOAD_CONFIGURATIONS_LOADED",
  UPLOAD_FILE_LOADING = "UPLOAD_FILE_LOADING",
  UPLOAD_FILE_LOADED = "UPLOAD_FILE_LOADED",
  UPLOADED_FILES_LOADING = "UPLOADED_FILES_LOADING",
  UPLOADED_FILES_LOADED = "UPLOADED_FILES_LOADED",
  DELETE_FILE_LOADING = "DELETE_FILE_LOADING",
  DELETE_FILE_LOADED = "DELETE_FILE_LOADED",
}